fix misc jax issues

+++ src/mlrp_course/trajopt/envs/double_integrator.py
@@ -142,7 +142,7 @@ class JaxDoubleIntegratorProblem(DoubleIntegratorProblem):
 
     @property
     def initial_state(self) -> TrajOptState:
-        return jnp.array(super().initial_state, dtype=jnp.float64)
+        return jnp.array(super().initial_state, dtype=jnp.float32)
 
     @staticmethod
     @jax.jit
@@ -158,14 +158,14 @@ class JaxDoubleIntegratorProblem(DoubleIntegratorProblem):
         next_x_dot = x_dot + u * dt
         next_x = x + next_x_dot * dt
 
-        return jnp.array([next_x, next_x_dot], dtype=jnp.float64)
+        return jnp.array([next_x, next_x_dot], dtype=jnp.float32)
 
     @staticmethod
     @jax.jit
     def _get_traj_cost(
-        xs: NDArray[jnp.float64],
-        x_dots: NDArray[jnp.float64],
-        actions: NDArray[jnp.float64],
+        xs: NDArray[jnp.float32],
+        x_dots: NDArray[jnp.float32],
+        actions: NDArray[jnp.float32],
         x_cost_weight: float,
         x_dot_cost_weight: float,
         torque_cost_weight: float,
